+++ src/main/kotlin/org/ironica/amatsukaze/playground/payloads/Payload.kt
@@ -8,26 +8,30 @@
  *
  */
 
-package org.ironica.amatsukaze.bridge
+package org.ironica.amatsukaze.playground.payloads
 
 import kotlinx.serialization.Serializable
-import org.ironica.amatsukaze.playground.*
-import org.ironica.amatsukaze.playground.data.Lock
+import org.ironica.amatsukaze.playground.Color
+import org.ironica.amatsukaze.playground.data.Grid
+import org.ironica.amatsukaze.playground.data.ItemLayout
 import org.ironica.amatsukaze.playground.data.Portal
-import org.ironica.amatsukaze.playground.data.Stair
-import org.ironica.amatsukaze.playground.enums.Block
-import org.ironica.amatsukaze.playground.enums.Item
 
 @Serializable
-data class Data(
-    val type: String,
-    val code: String,
-    val grid: List<List<Block>>,
-    val layout: List<List<Item>>,
+data class Payload(
+    val grid: Grid,
+    val itemLayout: ItemLayout,
     val colors: List<List<Color>>,
     val levels: List<List<Int>>,
+    val players: List<SerializedPlayer>,
     val portals: List<Portal>,
-    val locks: List<Lock>,
-    val stairs: List<Stair>,
-    val players: List<PlayerData>
-)
+    val consoleLog: String,
+    val special: String
+    ) {
+    constructor(
+        players: List<SerializedPlayer>,
+        portals: List<Portal>,
+        grid: SerializedPlayground,
+        consoleLog: String,
+        special: String
+    ): this(grid.grid, grid.itemLayout, grid.colors, grid.levels, players, portals, consoleLog, special)
+}
